chore: Rename write-to-db file
